+++ mod.ts
@@ -1,1 +1,21 @@
-export * from './exception.ts';
+export {
+    Exception,
+    ExceptionProps
+} from './lib/Exception.ts';
+
+export {
+    ifOk,
+    ifErr,
+    isOk,
+    isErr,
+    ResultOk,
+    ResultErr,
+    Result,
+    Ok,
+    Err
+} from './lib/Result.ts';
+
+export {
+    Optional,
+    InvalidOptionalGetException
+} from './lib/Optional.ts';