+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# exception
+# safety
+
+## v0.0.2
+    - Rename to safety
+    - Add Optional + Tests
+    - Add Result + Tests
 
 ## v0.0.1
     - Add ExceptionProps interface